feat: Level2 UI subview로 분리
1. Level2에서 구현한 이미지 + 정보 스택뷰를 따로 subview로 분리했습니다
