+++ suhyeon/HarryPotterBooks/HarryPotterBooks/Presentation/Home/HomeView.swift
@@ -23,51 +23,7 @@ class HomeView: UIView {
     lazy var seriesButton = SeriesButton(title: "1", widthHeightLength: widthHeightLength)
     
     // 책 정보 영역 스택뷰 (이미지 + 텍스트)
-    private let bookInfoStackView = UIStackView().then {
-        $0.axis = .horizontal
-        $0.alignment = .center
-        $0.spacing = 10
-        
-    }
-    
-    // 책 표지 이미지
-    private let bookImageView = UIImageView().then {
-        $0.contentMode = .scaleAspectFit
-    }
-    
-    // 책 정보 스택뷰 (이미지 x)
-    private let bookDescriptionStackView = UIStackView().then {
-        $0.axis = .vertical
-        $0.alignment = .leading
-        $0.spacing = 8
-    }
-    
-    // 책 제목
-    private let bookTitleLabel = UILabel().then {
-        $0.font = .systemFont(ofSize: 20, weight: .bold)
-        $0.textColor = .black
-        $0.numberOfLines = 0
-    }
-    
-    // 저자 스택뷰
-    private let authorStackView = InfoStackView(title: "Author",
-                                                titleFont:  .systemFont(ofSize: 16, weight: .bold),
-                                                titleColor: .black,
-                                                contentFont: .systemFont(ofSize: 18),
-                                                contentColor: .darkGray)
-    
-    // 출간일 스택뷰
-    private let releaseStackView = InfoStackView(title: "Released",
-                                                titleFont:  .systemFont(ofSize: 14, weight: .bold),
-                                                titleColor: .black,
-                                                contentFont: .systemFont(ofSize: 14),
-                                                contentColor: .gray)
-    // 페이지 스택뷰
-    private let pageStackView = InfoStackView(title: "Pages",
-                                                titleFont:  .systemFont(ofSize: 14, weight: .bold),
-                                                titleColor: .black,
-                                                contentFont: .systemFont(ofSize: 14),
-                                                contentColor: .gray)
+    private let bookDetailStackView = BookDetailStackView()
     
     override init(frame: CGRect) {
         super.init(frame: frame)
@@ -86,20 +42,10 @@ class HomeView: UIView {
         [
             mainTitleLabel,
             seriesButton,
-            bookInfoStackView
+            bookDetailStackView
         ].forEach { self.addSubview($0) } // HomeView
         
-        [
-            bookImageView,
-            bookDescriptionStackView
-        ].forEach { bookInfoStackView.addArrangedSubview($0) } // BookInfoStackView (이미지 + 책 설명)
-        
-        [
-            bookTitleLabel,
-            authorStackView,
-            releaseStackView,
-            pageStackView
-        ].forEach { bookDescriptionStackView.addArrangedSubview($0) } // 책 설명 스택뷰
+
     }
     
     // 제약조건 설정
@@ -120,25 +66,15 @@ class HomeView: UIView {
         }
         
         // 책 정보 스택 뷰
-        bookInfoStackView.snp.makeConstraints { make in
+        bookDetailStackView.snp.makeConstraints { make in
             make.top.equalTo(seriesButton.snp.bottom).offset(20)
             make.directionalHorizontalEdges.equalTo(safeAreaLayoutGuide).inset(5)
-        }
-        
-        // 책 이미지
-        bookImageView.snp.makeConstraints { make in
-            make.width.equalTo(100)
-            make.height.equalTo(bookImageView.snp.width).multipliedBy(1.5)
         }
     }
     
     // configuration
     public func configure(book: Book, index: Int) {
         mainTitleLabel.text = book.title
-        bookTitleLabel.text = book.title
-        bookImageView.image = UIImage(named: "harrypotter\(index)")
-        authorStackView.configure(content: book.author)
-        releaseStackView.configure(content: book.releaseDate)
-        pageStackView.configure(content: "\(book.pages)")
+        bookDetailStackView.configure(book: book, index: index)
     }
 }
